Move QueryExamples
