+++ client/branded/src/search-ui/components/QueryExamples.constants.ts
