+++ client/branded/src/search-ui/components/QueryExamples.module.scss
